requirements.txt
# API Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0

# ML/Data
lightgbm==4.3.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# MLflow
mlflow==2.10.2

# Hugging Face (for production model loading)
huggingface-hub>=0.20.0

# Streamlit (optional, for demo app)
streamlit==1.28.1
requests==2.31.0
plotly==5.17.0